requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
aiofiles==23.2.1
apscheduler==3.10.4